Gas price and config file
