+++ ui/src/constants/index.ts
@@ -1,3 +1,4 @@
 export const DraggableTypes = {
   PIECE: "piece",
 };
+export const WC_PROJECT_ID = import.meta.env.VITE_WC_PROJECT_ID;
